backend/requirements.txt
asgiref==3.7.2
Django>=3.2,<6.0
django-cors-headers==4.3.1
djangorestframework==3.15.0
sqlparse==0.4.4
djangorestframework-simplejwt==5.3.1